Add support for tex2D() to read from linear memory:
This is a CUDA-only feature. Instead of using tex1Dfetch, we can use tex2D to
read from a texture bound to linear memory. The address calculation is done in
hardware using tex2D. However, this seems to come at overhead in terms of
register usage. For the bilateral filter this results in a slowdown.
This feature is currently hardcoded (enabled/disabled) in the
HipaccKernelFeatures class.