Using Local Configurations per Algorithm
Instead of a global HIPACC_CONFIG flag in the Makefile for all the
kernels in the algorithm, a local config.json file can be used to
give kernel specific configurations. The original HIPACC_CONFIG now
serve as the default global configurations, which can be overwritten
by the local config file if desired.