Fixed OpenMP pragma generation with Clang 8
For pretty printing, omp pragmas now need an associated
statement, which is the actual for-loop. This statement must
be provided upon creating the pragma. The pragma itself
contains the for-loop now, so it is sufficient to insert the
pragma in the AST without the for-loop. Before, the pragma
AND the for-loop had to be added to the AST in the right
order.